// tooling/types.ts
export type BaseNetworkConfig = {
    chainId: number;
    url: string | undefined;

    // optionnal fields
    api_key?: string | undefined | null;
    forgeDeployExtraArgs?: string;
    profile?: string;
    forgeVerifyExtraArgs?: string;
    disableSourcify?: boolean;
    disableVerifyOnDeploy?: boolean;
    extra?: any;
    enumName?: string;
};

export type NetworkConfig = BaseNetworkConfig & {
    name: string;
    enumName: string;
    addresses: AddressSections;
};

export type AddressSections = {
    [key: string]: AddressSection;
};

export type AddressSection = {
    [key: string]: AddressEntry;
};

export type AddressEntry = {
    key: string;
    value: `0x${string}`;
};

export type BaseConfig = {
    deploymentFolder: string;
    defaultNetwork: string;
    networks: {
        [key: string]: BaseNetworkConfig;
    };
};